Fix broken tests
They rely on expression types that are not part of the IR
or are missing necessary casts.

(cherry picked from commit b6be3efb260b76644df1e6acbce7d88244a0fdfc)

+++ core/trino-main/src/test/java/io/trino/sql/planner/TestDomainTranslator.java
@@ -1440,7 +1440,7 @@ public class TestDomainTranslator
 
         assertPredicateIsAlwaysFalse(not(new InPredicate(
                 C_BIGINT.toSymbolReference(),
-                new InListExpression(ImmutableList.of(cast(toExpression(null, SMALLINT), BIGINT), toExpression(1L, SMALLINT))))));
+                new InListExpression(ImmutableList.of(cast(toExpression(null, SMALLINT), BIGINT), toExpression(1L, BIGINT))))));
     }
 
     @Test
